cloud-skills-main/Day-01/web-app/requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
boto3==1.28.40
botocore==1.31.40
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16